# document/api/api.yml
openapi: 3.0.3
info:
  title: PetStore API
  version: "0.1.1"
  description: |-
    ## PetStore OpenAPI 設計
    - バックエンド： Laravel
    - フロントエンド： Vue
  license:
    name: MIT
servers:
  - url: "http://localhost/api/"
    description: 開発サーバーURL
  - url: "https://product/api/"
    description: 本番サーバーURL
paths:
  /user:
    post:
      operationId: registerNewUser
      tags:
        - User
      summary: 新規ユーザー登録
      description: |-
        新規ユーザー登録
        ## Permission
          - None
        ## Validations
          - email: メールアドレスチェック、Usersテーブル重複チェック

          - name: 最大40文字
        ## Logic
          - 登録成功するとメール通知
          - 通知メールのURLに email_verity.verity_string
          - メールアドレスチェック後email_verified_at更新
          - email_verified_atがnullではない場合、認証済みユーザーと認識する。
      responses:
        default:
          description: その他エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 登録成功
          content:
            applicaiton/json:
              schema:
                properties:
                  id:
                    description: 新規登録ユーザーID
                    type: integer
                additionalProperties: false
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      requestBody:
        content:
          applicaiton/json:
            schema:
              $ref: "#/components/schemas/request_auth_register"
            example:
              name: "test"
              email: "test@test.com"
              password: "password"
      security:
        - apiKey: []
    get:
      operationId: getUser
      tags:
        - User
      summary: ユーザー情報取得
      description: |-
        ユーザー情報
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/user"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      security:
        - apiKey: []
        - bearer: []
    put:
      operationId: updateUser
      tags:
        - User
      summary: ユーザー情報更新
      description: |-
        ユーザー情報更新項目
        - ユーザー名更新
        - ステータス更新
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/response_ok"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      requestBody:
        content:
          applicaiton/json:
            schema:
              $ref: "#/components/schemas/user"
      security:
        - apiKey: []
        - bearer: []
    delete:
      operationId: deleteUser
      tags:
        - User
      summary: ユーザー削除
      description: |-
        ユーザー削除
        - softdelete
      responses:
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unexpected"
        "200":
          description: 成功
          content:
            applicaiton/json:
              schema:
                $ref: "#/components/schemas/response_ok"
        "401":
          description: <API-KEY> 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "403":
          description: セキュリティエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "404":
          description: Model not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error400"
        "500":
          description: サーバーエラー.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error500"
      security:
        - apiKey: []
        - bearer: []
...
components:
  schemas:
...
    user:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
    request_auth_register:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
...
    response_ok:
      required:
        - ok
      type: object
      properties:
        ok:
          type: boolean
...
    error400:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
    error500:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
    unexpected:
      type: object
      properties:
        status:
          type: string
          default: error
        message:
          type: string
        code:
          type: integer
  securitySchemes:
    apiKey:
      type: apiKey
      name: API-KEY
      in: header
    bearer:
      type: http
      scheme: bearer
      bearerFormat: bearer
security:
  - apiKey: []
  - bearer: []